requirements.txt
apptools==5.1.0
astropy==5.1.1
attrs==22.2.0
Bottleneck==1.3.8
certifi==2021.5.30
cftime==1.5.0
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
contourpy==1.2.0
csaps==1.1.0
cycler==0.10.0
dill==0.3.3
distlib==0.3.8
envisage==6.0.1
exceptiongroup==1.1.0
filelock==3.13.4
fonttools==4.33.3
h5py==3.4.0
idna==2.10
importlib-metadata==4.5.0
importlib-resources==5.1.4
iniconfig==2.0.0
joblib==1.3.2
kiwisolver==1.3.1
llvmlite==0.39.1
matplotlib==3.8.4
numba==0.56.4
numexpr==2.8.4
numpy==1.22.4
numpy-ext==0.9.9
orjson==3.8.6
packaging==24.1
pandas==2.2.2
patsy==0.5.1
Pillow==8.4.0
pipdeptree==2.23.0
platformdirs==4.1.0
pluggy==1.0.0
pooch==1.8.0
pycairo==1.27.0
pyerfa==2.0.0.1
pyevtk==1.6.0
pyface==7.3.0
Pygments==2.9.0
PyGObject==3.42.2
PyGObject-stubs==2.3.1
pyNFFT2==1.4.2
pyparsing==2.4.7
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
PySide2==5.15.2.1
pytest==7.2.2
python-dateutil==2.8.2
pytz==2021.3
pyvista==0.41.1
PyYAML==6.0.1
requests==2.26.0
rich==13.0.1
SciencePlots==2.0.1
scikit-guess==0.0.1a0
scikit-learn==1.4.1.post1
scipy==1.11.4
scooby==0.9.2
shiboken2==5.15.2.1
shiboken6==6.4.2
six==1.16.0
sklearn==0.0
sourceinspect==0.0.4
statsmodels==0.12.2
tables==3.7.0
taichi==1.3.0
threadpoolctl==2.1.0
tomli==2.0.1
tqdm==4.66.1
traits==6.2.0
traitsui==7.2.1
tzdata==2024.1
urllib3==1.26.5
virtualenv==20.25.1
vtk==9.3.0
zipp==3.4.1
